Rework mouse wheel emulation to operate on floats (SDL 3 preparation)

+++ include/mouse.h
@@ -92,7 +92,7 @@ void MOUSE_EventButton(const MouseButtonId button_id, const bool pressed);
 void MOUSE_EventButton(const MouseButtonId button_id, const bool pressed,
                        const MouseInterfaceId device_id);
 
-void MOUSE_EventWheel(const int16_t w_rel);
+void MOUSE_EventWheel(const float w_rel);
 void MOUSE_EventWheel(const int16_t w_rel, const MouseInterfaceId device_id);
 
 // Notify that guest OS is being booted, so that certain
